+++ tests/sdk/nodejs/examples/examples_test.go
@@ -20,6 +20,7 @@ import (
 	"path/filepath"
 	"regexp"
 	"sort"
+	"strings"
 	"testing"
 
 	"github.com/pulumi/pulumi-kubernetes/provider/v3/pkg/openapi"
@@ -139,7 +140,7 @@ func TestAccHelm(t *testing.T) {
 			ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
 				// Ensure that all `Services` have `status` marked as a `Secret`
 				for _, res := range stackInfo.Deployment.Resources {
-					if res.Type == tokens.Type("kubernetes:core/v1:Service") {
+					if res.Type == "kubernetes:core/v1:Service" {
 						spec, has := res.Outputs["status"]
 						assert.True(t, has)
 						specMap, is := spec.(map[string]interface{})
@@ -195,6 +196,18 @@ func TestAccHelmAllowCRDRendering(t *testing.T) {
 			ExtraRuntimeValidation: func(t *testing.T, stackInfo integration.RuntimeValidationStackInfo) {
 				assert.NotNil(t, stackInfo.Deployment)
 				assert.Equal(t, 8, len(stackInfo.Deployment.Resources))
+
+				for _, res := range stackInfo.Deployment.Resources {
+					if res.Type == "kubernetes:core/v1:Pod" {
+						annotations, ok := openapi.Pluck(res.Inputs, "metadata", "annotations")
+						if strings.Contains(res.ID.String(), "skip-crd") {
+							assert.False(t, ok)
+						} else {
+							assert.True(t, ok)
+							assert.Contains(t, annotations, "pulumi.com/skipAwait")
+						}
+					}
+				}
 			},
 		})
 	integration.ProgramTest(t, &test)